Restored strict and warnings pragmas to modules
There were a number of modules where the strict and warnings pragmas
had been removed. I restored these and fixed the warning message in
HMM/Profile (caused by PDL returning the string 'BAD' instead
of a number) by checking the string to make sure it is numeric, before
testing it.